+++ sdk/dotnet/Pulumi/Deployment/GrpcDispatcher.cs
@@ -1,0 +1,41 @@
+// Copyright 2016-2020, Pulumi Corporation
+
+using System.Threading.Tasks;
+using Grpc.Core;
+using Pulumirpc;
+
+namespace Pulumi
+{
+    internal class GrpcDispatcher : IRpcDispatcher
+    {
+        private readonly ResourceMonitor.ResourceMonitorClient _client;
+        private readonly Engine.EngineClient _engine;
+
+        public GrpcDispatcher(string engine, string monitor)
+        {
+            this._engine = new Engine.EngineClient(new Channel(engine, ChannelCredentials.Insecure));
+            this._client = new ResourceMonitor.ResourceMonitorClient(new Channel(monitor, ChannelCredentials.Insecure));
+        }
+
+        public async Task<InvokeResponse> InvokeAsync(InvokeRequest request)
+            => await this._client.InvokeAsync(request);
+
+        public async Task<ReadResourceResponse> ReadResourceAsync(Resource resource, ReadResourceRequest request)
+            => await this._client.ReadResourceAsync(request);
+
+        public async Task<RegisterResourceResponse> RegisterResourceAsync(Resource resource, RegisterResourceRequest request)
+            => await this._client.RegisterResourceAsync(request);
+
+        public async Task RegisterResourceOutputsAsync(RegisterResourceOutputsRequest request)
+            => await this._client.RegisterResourceOutputsAsync(request);
+
+        public async Task LogAsync(LogRequest request)
+            => await this._engine.LogAsync(request);
+
+        public async Task<SetRootResourceResponse> SetRootResourceAsync(SetRootResourceRequest request)
+            => await this._engine.SetRootResourceAsync(request);
+
+        public async Task<GetRootResourceResponse> GetRootResourceAsync(GetRootResourceRequest request)
+            => await this._engine.GetRootResourceAsync(request);
+    }
+}